refactor(k8s): split container build module up
