+++ core/src/plugins/kubernetes/container/handlers.ts
@@ -20,7 +20,7 @@ import { getContainerServiceStatus } from "./status"
 import { getTestResult } from "../test-results"
 import { ContainerModule } from "../../container/config"
 import { getTaskResult } from "../task-results"
-import { k8sBuildContainer, k8sGetContainerBuildStatus } from "./build"
+import { k8sBuildContainer, k8sGetContainerBuildStatus } from "./build/build"
 import { k8sPublishContainerModule } from "./publish"
 import { getPortForwardHandler } from "../port-forward"
 import { GetModuleOutputsParams } from "../../../types/plugin/module/getModuleOutputs"
