+++ core/src/plugins/openfaas/build.ts
@@ -12,7 +12,7 @@ import { KubernetesProvider } from "../kubernetes/config"
 import { dumpYaml } from "../../util/util"
 import { BuildModuleParams } from "../../types/plugin/module/build"
 import { containerHelpers } from "../container/helpers"
-import { k8sBuildContainer, k8sGetContainerBuildStatus } from "../kubernetes/container/build"
+import { k8sBuildContainer, k8sGetContainerBuildStatus } from "../kubernetes/container/build/build"
 import { GetBuildStatusParams } from "../../types/plugin/module/getBuildStatus"
 import { OpenFaasModule, getContainerModule, OpenFaasProvider } from "./config"
 import { LogEntry } from "../../logger/log-entry"
